feat: Integrate wallet library
