+++ package.json
@@ -353,6 +353,7 @@
     "@metamask/transaction-pay-controller": "^22.7.0",
     "@metamask/tron-wallet-snap": "^1.25.3",
     "@metamask/utils": "^11.11.0",
+    "@metamask/wallet": "^1.0.0",
     "@myx-trade/sdk": "^0.1.265",
     "@ngraveio/bc-ur": "^1.1.6",
     "@nktkas/hyperliquid": "^0.32.2",
